# docs/gateway_swagger.yaml
basePath: /$gateway/api/
definitions:
  curd.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  curd.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
...
  curd.ReplyData-types_Link:
    properties:
      data:
        $ref: '#/definitions/types.Link'
      error:
        type: string
    type: object
  curd.ReplyData-types_Serial:
    properties:
      data:
        $ref: '#/definitions/types.Serial'
      error:
        type: string
    type: object
...
  curd.ReplyList-types_Link:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Link'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  curd.ReplyList-types_Serial:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Serial'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
  types.Link:
    properties:
      created:
        description: 创建时间
        type: string
      description:
        description: 说明
        type: string
      disabled:
        type: boolean
      heartbeat:
        description: 心跳包
        type: string
      id:
        description: ID
        type: string
      name:
        description: 名称
        type: string
      poller_interval:
        description: 采集间隔
        type: integer
      poller_period:
        description: 采集周期
        type: integer
      protocol_name:
        description: 协议 rtu tcp parallel-tcp
        type: string
      protocol_options:
        additionalProperties: {}
        description: 协议参数
        type: object
      remote:
        description: 远程地址
        type: string
      running:
        type: boolean
      server_id:
        description: 服务器ID
        type: string
    type: object
  types.Serial:
    properties:
      baud_rate:
        description: 9600 115200
        type: integer
      created:
        description: 创建时间
        type: string
      data_bits:
        description: 5 6 7 8
        type: integer
      description:
        description: 说明
        type: string
      disabled:
        type: boolean
      heartbeat:
        description: 心跳包
        type: string
      id:
        description: ID
        type: string
      name:
        description: 名称
        type: string
      parity_mode:
        description: 0 1 2 NONE ODD EVEN
        type: integer
      poller_interval:
        description: 采集间隔
        type: integer
      poller_period:
        description: 采集周期
        type: integer
      port_name:
        description: port, e.g. COM1 "/dev/ttySerial1".
        type: string
      protocol_name:
        description: 协议 rtu tcp parallel-tcp
        type: string
      protocol_options:
        additionalProperties: {}
        description: 协议参数
        type: object
      retry_maximum:
        description: 最大次数
        type: integer
      retry_timeout:
        description: 重试时间
        type: integer
      running:
        type: boolean
      stop_bits:
        description: 1 2
        type: integer
    type: object
...
info:
  contact: {}
  description: API文档
  title: 物联大师网关接口文档
  version: 1.0 版本
paths:
...
  /link/{id}:
    get:
      consumes:
      - application/json
      description: 获取连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Link'
      summary: 获取连接
      tags:
      - link
    post:
      consumes:
      - application/json
      description: 修改连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      - description: 连接信息
        in: body
        name: link
        required: true
        schema:
          $ref: '#/definitions/types.Link'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Link'
      summary: 修改连接
      tags:
      - link
  /link/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Link'
      summary: 删除连接
      tags:
      - link
  /link/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Link'
      summary: 禁用连接
      tags:
      - link
  /link/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Link'
      summary: 启用连接
      tags:
      - link
  /link/{id}/stop:
    get:
      consumes:
      - application/json
      description: 停止连接
      parameters:
      - description: 连接ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Link'
      summary: 停止连接
      tags:
      - link
  /link/count:
    post:
      consumes:
      - application/json
      description: 查询连接数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询连接数量
      tags:
      - link
...
  /link/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入连接
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 导入连接
      tags:
      - link
  /link/list:
    get:
      consumes:
      - application/json
      description: 查询连接
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-types_Link'
      summary: 查询连接
      tags:
      - link
  /link/search:
    post:
      consumes:
      - application/json
      description: 查询连接
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-types_Link'
      summary: 查询连接
      tags:
      - link
  /serial/{id}:
    get:
      consumes:
      - application/json
      description: 获取串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Serial'
      summary: 获取串口
      tags:
      - serial
    post:
      consumes:
      - application/json
      description: 修改串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      - description: 串口信息
        in: body
        name: serial
        required: true
        schema:
          $ref: '#/definitions/types.Serial'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Serial'
      summary: 修改串口
      tags:
      - serial
  /serial/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Serial'
      summary: 删除串口
      tags:
      - serial
  /serial/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Serial'
      summary: 禁用串口
      tags:
      - serial
  /serial/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Serial'
      summary: 启用串口
      tags:
      - serial
  /serial/{id}/start:
    get:
      consumes:
      - application/json
      description: 启动串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Serial'
      summary: 启动串口
      tags:
      - serial
  /serial/{id}/stop:
    get:
      consumes:
      - application/json
      description: 停止串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Serial'
      summary: 停止串口
      tags:
      - serial
  /serial/count:
    post:
      consumes:
      - application/json
      description: 查询串口数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询串口数量
      tags:
      - serial
  /serial/create:
    post:
      consumes:
      - application/json
      description: 创建串口
      parameters:
      - description: 串口信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Serial'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-types_Serial'
      summary: 创建串口
      tags:
      - serial
...
  /serial/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入串口
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 导入串口
      tags:
      - product
  /serial/list:
    get:
      consumes:
      - application/json
      description: 查询串口
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-types_Serial'
      summary: 查询串口
      tags:
      - serial
...
  /serial/search:
    post:
      consumes:
      - application/json
      description: 查询串口
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-types_Serial'
      summary: 查询串口
      tags:
      - serial
...
swagger: "2.0"